+++ include/neuriplo/tasks/core/image_io.hpp
@@ -4,6 +4,7 @@
 
 namespace neuriplo_tasks {
 
+using vision::decodeImage;
 using vision::loadImage;
 using vision::saveImage;